sftp-mover/requirements.txt
google-cloud-bigquery=2.18.0
google-cloud-storage=1.40.0